+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- query: Fix query expression with keyword prefix like india or andrea
 - extractor: Except buggy Samsung images (by psdimon)
 - query: Fix query expressions with parenthesis
 - database: Unify city geo information from hamlet, village, town